refactoring: move parse node to individual file. Cleaned up the file names.
